requirements.txt
requests
scrapy
ipython
python-dotenv

scrapy-deltafetch
pandas
google-cloud-storage